rename root_client to root_provider in cloud_client/aws module
